Adjust metadata
